Combine DebugToolbarSystemChecksTestCase with ChecksTestCase
Both classes aim to test system checks.

Using a separate module is nice to avoid test_integrations.py from
becoming too big.
