+++ tests/test_integration.py
@@ -3,14 +3,12 @@ import re
 import unittest
 
 import html5lib
-from django.conf import settings
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from django.core import signing
-from django.core.checks import Warning, run_checks
 from django.db import connection
 from django.http import HttpResponse
 from django.template.loader import get_template
-from django.test import RequestFactory, SimpleTestCase
+from django.test import RequestFactory
 from django.test.utils import override_settings
 
 from debug_toolbar.middleware import DebugToolbarMiddleware, show_toolbar
@@ -28,7 +26,6 @@ except ImportError:
     webdriver = None
 
 
-PATH_DOES_NOT_EXIST = os.path.join(settings.BASE_DIR, "tests", "invalid_static")
 rf = RequestFactory()
 
 
@@ -440,104 +437,4 @@ class DebugToolbarLiveTestCase(StaticLiveServerTestCase):
             lambda selenium: self.selenium.find_element_by_css_selector(
                 "#djDebugWindow code"
             )
-        )
-
-
-@override_settings(DEBUG=True)
-class DebugToolbarSystemChecksTestCase(SimpleTestCase):
-    @override_settings(
-        MIDDLEWARE=[
-            "django.contrib.messages.middleware.MessageMiddleware",
-            "django.contrib.sessions.middleware.SessionMiddleware",
-            "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "django.middleware.gzip.GZipMiddleware",
-            "debug_toolbar.middleware.DebugToolbarMiddleware",
-        ]
-    )
-    def test_check_good_configuration(self):
-        messages = run_checks()
-        self.assertEqual(messages, [])
-
-    @override_settings(
-        MIDDLEWARE=[
-            "django.contrib.messages.middleware.MessageMiddleware",
-            "django.contrib.sessions.middleware.SessionMiddleware",
-            "django.contrib.auth.middleware.AuthenticationMiddleware",
-        ]
-    )
-    def test_check_missing_middleware_error(self):
-        messages = run_checks()
-        self.assertEqual(
-            messages,
-            [
-                Warning(
-                    "debug_toolbar.middleware.DebugToolbarMiddleware is "
-                    "missing from MIDDLEWARE.",
-                    hint="Add debug_toolbar.middleware.DebugToolbarMiddleware "
-                    "to MIDDLEWARE.",
-                    id="debug_toolbar.W001",
-                )
-            ],
-        )
-
-    @override_settings(
-        MIDDLEWARE=[
-            "django.contrib.messages.middleware.MessageMiddleware",
-            "django.contrib.sessions.middleware.SessionMiddleware",
-            "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "debug_toolbar.middleware.DebugToolbarMiddleware",
-            "django.middleware.gzip.GZipMiddleware",
-        ]
-    )
-    def test_check_gzip_middleware_error(self):
-        messages = run_checks()
-        self.assertEqual(
-            messages,
-            [
-                Warning(
-                    "debug_toolbar.middleware.DebugToolbarMiddleware occurs "
-                    "before django.middleware.gzip.GZipMiddleware in "
-                    "MIDDLEWARE.",
-                    hint="Move debug_toolbar.middleware.DebugToolbarMiddleware "
-                    "to after django.middleware.gzip.GZipMiddleware in "
-                    "MIDDLEWARE.",
-                    id="debug_toolbar.W003",
-                )
-            ],
-        )
-
-    @override_settings(
-        MIDDLEWARE_CLASSES=[
-            "django.contrib.messages.middleware.MessageMiddleware",
-            "django.contrib.sessions.middleware.SessionMiddleware",
-            "django.contrib.auth.middleware.AuthenticationMiddleware",
-            "django.middleware.gzip.GZipMiddleware",
-            "debug_toolbar.middleware.DebugToolbarMiddleware",
-        ]
-    )
-    def test_check_middleware_classes_error(self):
-        messages = run_checks()
-        self.assertIn(
-            Warning(
-                "debug_toolbar is incompatible with MIDDLEWARE_CLASSES setting.",
-                hint="Use MIDDLEWARE instead of MIDDLEWARE_CLASSES",
-                id="debug_toolbar.W004",
-            ),
-            messages,
-        )
-
-    @override_settings(
-        STATICFILES_DIRS=[PATH_DOES_NOT_EXIST],
-    )
-    def test_panel_check_errors(self):
-        messages = run_checks()
-        self.assertEqual(
-            messages,
-            [
-                Warning(
-                    "debug_toolbar requires the STATICFILES_DIRS directories to exist.",
-                    hint="Running manage.py collectstatic may help uncover the issue.",
-                    id="debug_toolbar.staticfiles.W001",
-                )
-            ],
         )
